src/dbcontrolservice/requirements.txt
Flask
mysql-connector-python
grpcio
grpcio-tools
